Moved external interface for EpochStateMachine in epochs package
