+++ state/protocol/protocol_state/state/mutator_test.go
@@ -3,6 +3,7 @@ package state
 import (
 	"errors"
 	"fmt"
+	"github.com/onflow/flow-go/state/protocol/protocol_state/epochs"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -14,8 +15,7 @@ import (
 	"github.com/onflow/flow-go/module/irrecoverable"
 	"github.com/onflow/flow-go/state/protocol"
 	protocolmock "github.com/onflow/flow-go/state/protocol/mock"
-	"github.com/onflow/flow-go/state/protocol/protocol_state"
-	protocolstatemock "github.com/onflow/flow-go/state/protocol/protocol_state/mock"
+	epochsmock "github.com/onflow/flow-go/state/protocol/protocol_state/epochs/mock"
 	"github.com/onflow/flow-go/storage/badger/transaction"
 	storagemock "github.com/onflow/flow-go/storage/mock"
 	"github.com/onflow/flow-go/utils/rand"
@@ -35,7 +35,7 @@ type StateMutatorSuite struct {
 	commitsDB       *storagemock.EpochCommits
 	globalParams    *protocolmock.GlobalParams
 	parentState     *flow.RichProtocolStateEntry
-	stateMachine    *protocolstatemock.EpochStateMachine
+	stateMachine    *epochsmock.EpochStateMachine
 	candidateView   uint64
 
 	mutator *stateMutator
@@ -51,7 +51,7 @@ func (s *StateMutatorSuite) SetupTest() {
 	s.globalParams.On("EpochCommitSafetyThreshold").Return(uint64(1_000))
 	s.parentState = unittest.ProtocolStateFixture()
 	s.candidateView = s.parentState.CurrentEpochSetup.FirstView + 1
-	s.stateMachine = protocolstatemock.NewEpochStateMachine(s.T())
+	s.stateMachine = epochsmock.NewEpochStateMachine(s.T())
 
 	var err error
 	s.mutator, err = newStateMutator(
@@ -62,10 +62,10 @@ func (s *StateMutatorSuite) SetupTest() {
 		s.globalParams,
 		s.candidateView,
 		s.parentState,
-		func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+		func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 			return s.stateMachine, nil
 		},
-		func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+		func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 			require.Fail(s.T(), "entering epoch fallback is not expected")
 			return nil, fmt.Errorf("not expecting epoch fallback")
 		},
@@ -149,11 +149,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			expectedConstructorCalled := false
 			s.candidateView = s.parentState.CurrentEpochSetup.FirstView + 1
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					expectedConstructorCalled = true // expect happy-path constructor
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
@@ -168,11 +168,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			expectedConstructorCalled := false
 			s.candidateView = s.parentState.CurrentEpochSetup.FinalView - 1
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					expectedConstructorCalled = true // expect epoch-fallback state machine
 					return s.stateMachine, nil
 				},
@@ -193,11 +193,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			expectedConstructorCalled := false
 			s.candidateView = s.parentState.CurrentEpochSetup.FirstView + 1
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					expectedConstructorCalled = true // expect happy-path constructor
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
@@ -212,11 +212,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			expectedConstructorCalled := false
 			s.candidateView = s.parentState.CurrentEpochSetup.FinalView - 1
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					expectedConstructorCalled = true // expect epoch-fallback state machine
 					return s.stateMachine, nil
 				},
@@ -234,11 +234,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			expectedConstructorCalled := false
 			s.candidateView = s.parentState.CurrentEpochSetup.FirstView + 1
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					expectedConstructorCalled = true // expect happy-path constructor
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
@@ -253,11 +253,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			expectedConstructorCalled := false
 			s.candidateView = s.parentState.CurrentEpochSetup.FinalView - 1
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					expectedConstructorCalled = true // expect happy-path constructor
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
@@ -274,10 +274,10 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 		s.Run("happy-path", func() {
 			exception := irrecoverable.NewExceptionf("exception")
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					return nil, exception
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
@@ -290,11 +290,11 @@ func (s *StateMutatorSuite) TestStateMutator_Constructor() {
 			s.parentState.InvalidEpochTransitionAttempted = true // ensure we use epoch-fallback state machine
 			exception := irrecoverable.NewExceptionf("exception")
 			mutator, err := newStateMutator(s.headersDB, s.resultsDB, s.setupsDB, s.commitsDB, s.globalParams, s.candidateView, s.parentState,
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					s.T().Fail()
 					return s.stateMachine, nil
 				},
-				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+				func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 					return nil, exception
 				},
 			)
@@ -317,11 +317,11 @@ func (s *StateMutatorSuite) TestApplyServiceEvents_InvalidEpochSetup() {
 			s.globalParams,
 			s.candidateView,
 			s.parentState,
-			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 				return s.stateMachine, nil
 			},
-			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
-				fallbackStateMachine := protocolstatemock.NewEpochStateMachine(s.T())
+			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
+				fallbackStateMachine := epochsmock.NewEpochStateMachine(s.T())
 				fallbackStateMachine.On("ProcessEpochSetup", mock.Anything).Return(false, nil)
 				return fallbackStateMachine, nil
 			},
@@ -380,11 +380,11 @@ func (s *StateMutatorSuite) TestApplyServiceEvents_InvalidEpochCommit() {
 			s.globalParams,
 			s.candidateView,
 			s.parentState,
-			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
+			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
 				return s.stateMachine, nil
 			},
-			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (protocol_state.EpochStateMachine, error) {
-				fallbackStateMachine := protocolstatemock.NewEpochStateMachine(s.T())
+			func(candidateView uint64, parentState *flow.RichProtocolStateEntry) (epochs.EpochStateMachine, error) {
+				fallbackStateMachine := epochsmock.NewEpochStateMachine(s.T())
 				fallbackStateMachine.On("ProcessEpochCommit", mock.Anything).Return(false, nil)
 				return fallbackStateMachine, nil
 			},
